Third pass, bumped the minimal CMake version up to 3.4

+++ src/wrapped/wrappedlib_init.h
@@ -194,10 +194,10 @@ int FUNC(_fini)(library_t* lib)
     return 1;
 }
 
-int FUNC(_get)(library_t* lib, const char* name, uintptr_t *offs, uint32_t *sz)
+int FUNC(_get)(library_t* lib, const char* name, uintptr_t *offs, uintptr_t *sz)
 {
     uintptr_t addr = 0;
-    uint32_t size = 0;
+    uintptr_t size = 0;
 #ifdef CUSTOM_FAIL
     void* symbol = NULL;
 #endif
@@ -217,10 +217,10 @@ int FUNC(_get)(library_t* lib, const char* name, uintptr_t *offs, uint32_t *sz)
     return 1;
 }
 
-int FUNC(_getnoweak)(library_t* lib, const char* name, uintptr_t *offs, uint32_t *sz)
+int FUNC(_getnoweak)(library_t* lib, const char* name, uintptr_t *offs, uintptr_t *sz)
 {
     uintptr_t addr = 0;
-    uint32_t size = 0;
+    uintptr_t size = 0;
 #ifdef CUSTOM_FAIL
     void* symbol = NULL;
 #endif
